// src/interfaces/interpreter-commands.interface.ts
export enum INTERPRETER_COMMANDS {
  FORWARD = 'forward',
  BACKWARD = 'backward',
  TURN_LEFT = 'turnleft',
  TURN_RIGHT = 'turnright',
  DIRECTION = 'direction',
  CENTER = 'center',
  GO = 'go',
  GOX = 'gox',
  GOY = 'goy',
  PEN_UP = 'penup',
  PEN_DOWN = 'pendown',
  PEN_WIDTH = 'penwidth',
  PEN_COLOR = 'pencolor'
}